rename #20 - 컨벤션에 맞게 supebase user api 파일 이름 수정
